fix(review): move Severity uppercase regression test to tests/core/test_types.py
test_severity_enum_not_equal_to_uppercase_string tests a core-type StrEnum
invariant, not recipe behaviour. Move it to tests/core/test_types.py alongside
test_severity_has_ok_member where other Severity enum semantics are already
verified. Remove the now-unused Severity import from test_planner_recipe.py.
