+++ tests/recipe/test_planner_recipe.py
@@ -4,7 +4,6 @@ import importlib
 
 import pytest
 
-from autoskillit.core.types import Severity
 from autoskillit.recipe.io import builtin_recipes_dir, load_recipe
 from autoskillit.recipe.validator import run_semantic_rules
 from tests.recipe.conftest import assert_no_rule_errors
@@ -103,18 +102,6 @@ def test_planner_recipe_validate_routes_to_refine_on_fail(planner_recipe):
 def test_planner_recipe_validation_has_no_errors(planner_recipe):
     findings = run_semantic_rules(planner_recipe)
     assert_no_rule_errors(findings, context="planner recipe")
-
-
-def test_severity_enum_not_equal_to_uppercase_string():
-    """Regression: StrEnum values are lowercase; uppercase comparison is always False.
-
-    This bug made ``test_planner_recipe_validation_has_no_errors`` vacuous:
-    ``f.severity == "ERROR"`` always returned False because Severity.ERROR.value
-    is ``"error"`` (lowercase), not ``"ERROR"``.
-    """
-    assert Severity.ERROR != "ERROR"
-    assert Severity.ERROR == "error"
-    assert Severity.ERROR == Severity.ERROR
 
 
 def test_planner_recipe_extract_domain_uses_positional_args(planner_recipe):
